Imgui docking (#2)
* imgui working

@todo check why I need to include in ImGuiLayer.cpp :

#include "backends/imgui_impl_glfw.cpp"
#include "backends/imgui_impl_opengl3.cpp"

+++ Rebelle/src/Rebelle/Events/MouseEvent.h
@@ -57,7 +57,7 @@ namespace Rebelle {
 
 	protected:
 		MouseButtonEvent(int button) : m_Button(button) {}
-		float m_Button;
+		int m_Button;
 	};
 
 	class REBELLE_API MouseButtonPressedEvent : public MouseButtonEvent
